Fix for user content removal

+++ service-map/src/main/java/fi/nls/oskari/map/analysis/service/UserContentAnalysisService.java
@@ -1,16 +1,28 @@
 package fi.nls.oskari.map.analysis.service;
 
 import fi.nls.oskari.annotation.Oskari;
+import fi.nls.oskari.db.DatasourceHelper;
 import fi.nls.oskari.domain.User;
 import fi.nls.oskari.service.ServiceException;
 import fi.nls.oskari.service.db.UserContentService;
 
-@Oskari
+@Oskari("analysis")
 public class UserContentAnalysisService extends UserContentService {
 
-    private AnalysisDbServiceIbatisImpl analysisService = new AnalysisDbServiceIbatisImpl();
+    private AnalysisDbServiceIbatisImpl analysisService = null;
+
+    @Override
+    public void init() {
+        super.init();
+        if(DatasourceHelper.isModuleEnabled(getName())) {
+            analysisService = new AnalysisDbServiceIbatisImpl();
+        }
+    }
 
     public void deleteUserContent(User user) throws ServiceException {
+        if(!DatasourceHelper.isModuleEnabled(getName())) {
+            return;
+        }
         analysisService.deleteAnalysisByUid(user.getUuid());
     }
 }
